KM-26 refactor: MemberDeleter에서 OAuth2 연결 제거에 대한 책임을 인터페이스로 분리
