+++ src/main/java/kkomo/member/domain/OAuth2Unlinker.java
@@ -1,0 +1,6 @@
+package kkomo.member.domain;
+
+public interface OAuth2Unlinker {
+
+    void unlink(String accessToken);
+}
